crime python code/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.3
docopt==0.6.2
face-recognition @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_1619808233029/work
face-recognition-models @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_models_1595707436534/work
idna==3.7
mkl-fft==1.0.15
mkl-random==1.0.2
mkl-service==2.3.0
mysql-connector-python==8.0.18
numpy==1.19.5
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.9.0.80
Pillow==4.2.1
pipreqs==0.4.11
protobuf==3.6.0
requests==2.27.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
urllib3==1.26.18
wincertstore==0.2
yarg==0.1.9
